+++ Mercury.Reservations/src/Mercury.Reservations.Service/Dtos/RoomDto.cs
@@ -5,7 +5,16 @@ using Mercury.Common;
 
 namespace Mercury.Reservations.Service.Dtos
 {
-    public record RoomDto(Guid Id, string Title, string Description, int NumberOfTickets, DateTimeOffset CreatedAt, DateTimeOffset UpdatedAt, IEnumerable<TicketDto> Tickets);
+    public record RoomDto(
+        Guid Id,
+        string Title,
+        string Description,
+        int NumberOfTickets,
+        DateTimeOffset CreatedAt,
+        DateTimeOffset UpdatedAt,
+        IEnumerable<TicketDto> Tickets,
+        DateTimeOffset ExpiresAt
+    );
 
     public record TicketDto(Guid Id, string Folio, string Status, decimal Price);
 
@@ -13,6 +22,7 @@ namespace Mercury.Reservations.Service.Dtos
         [RequireNonDefault] Guid Id,
         [Required] string Title,
         string Description,
-        [Range(1, int.MaxValue)] int NumberOfTickets
+        [Range(1, int.MaxValue)] int NumberOfTickets,
+        [DateTimeValid] DateTimeOffset ExpiresAt
     );
 }